requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.11.1
boto3==1.28.3
botocore==1.31.3
certifi==2023.5.7
charset-normalizer==3.2.0
dj-database-url==2.0.0
Django==4.2.3
django-bootstrap5==21.3
django-heroku==0.3.1
django-storages==1.13.2
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
openai==0.27.8
Pillow==10.0.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==1.26.16
whitenoise==6.5.0
yarl==1.9.2
